// apps/www/src/generated/eslint-versions.ts
// This file is generated automatically
// DO NOT EDIT IT MANUALLY
import { type ESLintVersions } from "@/types/eslint";

export const eslintVersions: ESLintVersions = [
  {
    configs: ["base", "jest", "next", "prettier", "react", "tailwindcss"],
    version: "3.1.1",
  },
  {
    configs: ["base", "jest", "next", "prettier", "react", "tailwindcss"],
    version: "3.1.0",
  },
  {
    configs: ["base", "jest", "next", "prettier", "react", "tailwindcss"],
    version: "3.0.0",
  },
  { configs: ["base", "jest", "next", "prettier", "react"], version: "2.0.0" },
];